Change deployment strategy (#58)
This is a change to the deployment strategy and workflow. Intead of
deploying to the equivalent branch that code is merged into we group
deployments into two buckets, dev and prod. When code is merged into
the development branch travis will deploy to the dev account (i.e.
scipooldev). When code is merged to the prod branch travis will deploy,
in parallel, to all prod accounts (i.e. scipoolprod, strides, etc..).

The goal is to make it easier and faster to deploy to production accounts.