[SC-346] Download file containing Synapse token
The login app currently displays the requested Synapse token in the browser.
The problem with this is that the browser can apply language translation to
the token which modifies it and makes it invalid.

To avoid the problem described we are changing the login app to download a
file containing the token string instead of displaying the token directly
in the browser.