[SC-392] Update CI
Fix bit rot, bionic contains python 3.6 which has been deprecated and
removed from travis.  The build reports the following failure..

```
/usr/lib/python3/dist-packages/OpenSSL/crypto.py:12: CryptographyDeprecationWarning:
Python 3.6 is no longer supported by the Python core team. Therefore, support for
it is deprecated in cryptography and will be removed in a future release.
```

Attempt to fix by updating the distro.  Also updated simplified the CI config
a bit with better workflow and approach to installing AWS dependencies.